Refactor nestings for scast subcommands
